+++ src/main/java/hr/dlatecki/algorithms/gen_alg/operators/abstracts/AbstractSelectionOperator.java
@@ -10,7 +10,7 @@ import hr.dlatecki.algorithms.gen_alg.population.interfaces.IChromosome;
  * this class instead of directly implementing <code>ISelectionOperator</code> interface.
  * 
  * @author Domagoj Latečki
- * @version 1.0
+ * @version 1.1
  * @since 1.8
  * @param <C> Type of chromosome which will be used in the selection operator.
  * @see IChromosome
@@ -22,7 +22,7 @@ public abstract class AbstractSelectionOperator<C extends IChromosome> extends A
     /**
      * Defines a number of chromosomes to be selected in this selection operator.
      */
-    protected int selectionSize;
+    private int selectionSize;
     
     /**
      * Constructs a selection operator with provided <code>Random</code> object and selection size. Selection size must
@@ -30,14 +30,18 @@ public abstract class AbstractSelectionOperator<C extends IChromosome> extends A
      * 
      * @param rand object used to generate random numbers.
      * @param selectionSize number of chromosomes to select. Must be a number greater than or equal to 1.
+     * @throws IllegalArgumentException thrown if provided selection size is less than 1.
      */
     public AbstractSelectionOperator(Random rand, int selectionSize) {
         super(rand);
-        this.selectionSize = selectionSize;
+        
+        setSelectionSize(selectionSize);
     }
     
     @Override
     public void setSelectionSize(int size) {
+        
+        checkSelectionSize(size);
         
         selectionSize = size;
     }
@@ -47,4 +51,41 @@ public abstract class AbstractSelectionOperator<C extends IChromosome> extends A
         
         return select(pool, selectionSize);
     }
+    
+    @Override
+    public SortedSet<C> select(SortedSet<C> pool, int size) {
+        
+        checkSelectionSize(size);
+        
+        return performSelection(pool, size);
+    }
+    
+    /**
+     * Checks if provided selection size is greater or equal to 1.
+     * 
+     * @param size selection size.
+     * @throws IllegalArgumentException thrown if provided selection size is less than 1.
+     */
+    private void checkSelectionSize(int size) {
+        
+        if (size < 1) {
+            throw new IllegalArgumentException("Selection size cannot be less than 1.");
+        }
+    }
+    
+    /**
+     * Performs the selection. Chromosomes are selected based on their fitness values, with chromosomes with higher
+     * fitness having higher chance of being selected. The provided set of chromosomes is expected to be sorted by
+     * fitness in descending order. Number of selected chromosomes is defined by the second argument, which will always
+     * be greater than or equal to 1 (this is insured in <code>AbstractSelectionOperator</code>) before calling this
+     * method. This method is not supposed to be invoked externally. Instead, the {@link #select(SortedSet)} or
+     * {@link #select(SortedSet, int)} is invoked externally, and then this method is invoked by either of them after
+     * the size is confirmed to be correct.
+     * 
+     * @param pool set from which chromosomes will be selected based on their fitness.
+     * @param size number of chromosomes to select. Guaranteed to be greater than or equal to 1.
+     * @return Set which contains selected chromosomes. The set is sorted by fitness of the chromosomes, in descending
+     *         order.
+     */
+    protected abstract SortedSet<C> performSelection(SortedSet<C> pool, int size);
 }